+++ src/impls/android.rs
@@ -25,16 +25,26 @@ impl<R: Runtime> AndroidFsImpl<R> {
 }
 
 
-macro_rules! impl_serde {
+macro_rules! impl_se {
     (struct $struct_ident:ident $(< $lifetime:lifetime >)? { $( $name:ident: $ty:ty ),* $(,)? }) => {
-        #[derive(Serialize, Deserialize)]
+        #[derive(Serialize)]
+        #[serde(rename_all = "camelCase")]
+        struct $struct_ident $(< $lifetime >)? {
+            $($name: $ty,)*
+        }
+    };
+}
+
+macro_rules! impl_de {
+    (struct $struct_ident:ident $(< $lifetime:lifetime >)? { $( $name:ident: $ty:ty ),* $(,)? }) => {
+        #[derive(Deserialize)]
         #[serde(rename_all = "camelCase")]
         struct $struct_ident $(< $lifetime >)? {
             $($name: $ty,)*
         }
     };
     (struct $struct_ident:ident $(;)?) => {
-        #[derive(Serialize, Deserialize)]
+        #[derive(Deserialize)]
         struct $struct_ident;
     };
 }
@@ -42,10 +52,8 @@ macro_rules! impl_serde {
 impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
 
     fn get_name(&self, uri: &FileUri) -> crate::Result<String> {
-        impl_serde!(struct Req { uri: FileUri });
-        impl_serde!(struct Res { name: String });
-
-        let uri = uri.clone();
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res { name: String });
 
         self.api
             .run_mobile_plugin::<Res>("getName", Req { uri })
@@ -54,10 +62,8 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
 
     fn get_mime_type(&self, uri: &FileUri) -> crate::Result<Option<String>> {
-        impl_serde!(struct Req { uri: FileUri});
-        impl_serde!(struct Res { value: Option<String> });
-
-        let uri = uri.clone();
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res { value: Option<String> });
 
         self.api
             .run_mobile_plugin::<Res>("getMimeType", Req { uri })
@@ -66,10 +72,9 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
 
     fn open_file(&self, uri: &FileUri, mode: FileAccessMode) -> crate::Result<std::fs::File> {
-        impl_serde!(struct Req<'a> { uri: FileUri, mode: &'a str });
-        impl_serde!(struct Res { fd: std::os::fd::RawFd });
+        impl_se!(struct Req<'a> { uri: &'a FileUri, mode: &'a str });
+        impl_de!(struct Res { fd: std::os::fd::RawFd });
     
-        let uri = uri.clone();
         let mode = match mode {
             FileAccessMode::Read => "r",
             FileAccessMode::Write => "w",
@@ -95,11 +100,12 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
         multiple: bool,
     ) -> crate::Result<Vec<FileUri>> {
 		
-        impl_serde!(struct Req { mime_types: Vec<String>, multiple: bool, initial_location: Option<FileUri> });
-        impl_serde!(struct Res { uris: Vec<FileUri> });
-    
-        let initial_location = initial_location.map(Clone::clone);
-        let mime_types = mime_types.iter().map(|s| s.to_string()).collect();
+        impl_se!(struct Req<'a> { 
+            mime_types: &'a [&'a str],
+            multiple: bool,
+            initial_location: Option<&'a FileUri>
+        });
+        impl_de!(struct Res { uris: Vec<FileUri> });
 
         let _guard = self.intent_lock.lock();
         self.api
@@ -114,8 +120,8 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
         multiple: bool,
     ) -> crate::Result<Vec<FileUri>> {
 		
-        impl_serde!(struct Req { multiple: bool, target: VisualMediaTarget });
-        impl_serde!(struct Res { uris: Vec<FileUri> });
+        impl_se!(struct Req { multiple: bool, target: VisualMediaTarget });
+        impl_de!(struct Res { uris: Vec<FileUri> });
     
         let _guard = self.intent_lock.lock();
         self.api
@@ -131,12 +137,17 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
         mime_type: Option<&str>,
     ) -> crate::Result<Option<FileUri>> {
 
-        impl_serde!(struct Req<'a> { initial_file_name: &'a str, mime_type: &'a str, initial_location: Option<FileUri> });
-        impl_serde!(struct Res { uri: Option<FileUri> });
+        impl_se!(struct Req<'a> {
+            initial_file_name: &'a str, 
+            mime_type: &'a str, 
+            initial_location: Option<&'a FileUri> 
+        });
+        impl_de!(struct Res { uri: Option<FileUri> });
 
-        let initial_location = initial_location.map(Clone::clone);
         let initial_file_name = initial_file_name.as_ref();
-        let mime_type = mime_type.as_ref().map(|s| s.as_ref()).unwrap_or("application/octet-stream");
+        let mime_type = mime_type.as_ref()
+            .map(AsRef::as_ref)
+            .unwrap_or("application/octet-stream");
     
         let _guard = self.intent_lock.lock();
         self.api
@@ -150,10 +161,8 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
         initial_location: Option<&FileUri>,
     ) -> crate::Result<Option<FileUri>> {
         
-        impl_serde!(struct Req { initial_location: Option<FileUri> });
-        impl_serde!(struct Res { uri: Option<FileUri> });
-    
-        let initial_location = initial_location.map(Clone::clone);
+        impl_se!(struct Req<'a> { initial_location: Option<&'a FileUri> });
+        impl_de!(struct Res { uri: Option<FileUri> });
 
         let _guard = self.intent_lock.lock();
         self.api
@@ -161,12 +170,50 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
             .map(|v| v.uri)
             .map_err(Into::into)
     }
+
+    fn show_share_file_dialog(&self, uri: &FileUri) -> crate::Result<()> {
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res;);
+
+        self.api
+            .run_mobile_plugin::<Res>("shareFile", Req { uri })
+            .map(|_| ())
+            .map_err(Into::into)
+    }
+
+    fn can_share_file(&self, uri: &FileUri) -> crate::Result<bool> {
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res { value: bool });
+
+        self.api
+            .run_mobile_plugin::<Res>("canShareFile", Req { uri })
+            .map(|v| v.value)
+            .map_err(Into::into)
+    }
+
+    fn show_view_file_dialog(&self, uri: &FileUri) -> crate::Result<()> {
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res;);
+    
+        self.api
+            .run_mobile_plugin::<Res>("viewFile", Req { uri })
+            .map(|_| ())
+            .map_err(Into::into)
+    }
+
+    fn can_view_file(&self, uri: &FileUri) -> crate::Result<bool> {
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res { value: bool });
+
+        self.api
+            .run_mobile_plugin::<Res>("canViewFile", Req { uri })
+            .map(|v| v.value)
+            .map_err(Into::into)
+    }
     
     fn remove_file(&self, uri: &FileUri) -> crate::Result<()> {
-        impl_serde!(struct Req { uri: FileUri });
-        impl_serde!(struct Res;);
-
-        let uri = uri.clone();
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res;);
     
         self.api
             .run_mobile_plugin::<Res>("delete", Req { uri })
@@ -185,11 +232,10 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
         mime_type: Option<&str>
     ) -> crate::Result<FileUri> {
 
-        impl_serde!(struct Req<'a> { dir: FileUri, mime_type: &'a str, relative_path: &'a str });
+        impl_se!(struct Req<'a> { dir: &'a FileUri, mime_type: &'a str, relative_path: &'a str });
         
         let relative_path = relative_path.as_ref();
         let mime_type = mime_type.unwrap_or("application/octet-stream");
-        let dir = dir.clone();
 
         self.api
             .run_mobile_plugin::<FileUri>("createFile", Req { dir, mime_type, relative_path })
@@ -197,11 +243,8 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
 
     fn copy_via_kotlin(&self, src: &FileUri, dest: &FileUri) -> crate::Result<()> {
-        impl_serde!(struct Req { src: FileUri, dest: FileUri });
-        impl_serde!(struct Res;);
-
-        let src = src.clone();
-        let dest = dest.clone();
+        impl_se!(struct Req<'a> { src: &'a FileUri, dest: &'a FileUri });
+        impl_de!(struct Res;);
 
         self.api
             .run_mobile_plugin::<Res>("copyFile", Req { src, dest })
@@ -210,11 +253,9 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
     
     fn read_dir(&self, uri: &FileUri) -> crate::Result<impl Iterator<Item = Entry>> {
-        impl_serde!(struct Req { uri: FileUri });
-        impl_serde!(struct Obj { name: String, uri: FileUri, last_modified: i64, byte_size: i64, mime_type: Option<String> });
-        impl_serde!(struct Res { entries: Vec<Obj> });
-
-        let uri = uri.clone();
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Obj { name: String, uri: FileUri, last_modified: i64, byte_size: i64, mime_type: Option<String> });
+        impl_de!(struct Res { entries: Vec<Obj> });
     
         self.api
             .run_mobile_plugin::<Res>("readDir", Req { uri })
@@ -237,10 +278,8 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
 
     fn take_persistable_uri_permission(&self, uri: &FileUri) -> crate::Result<()> {
-        impl_serde!(struct Req { uri: FileUri });
-        impl_serde!(struct Res;);
-
-        let uri = uri.clone();
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res;);
 
         self.api
             .run_mobile_plugin::<Res>("takePersistableUriPermission", Req { uri })
@@ -249,10 +288,8 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
 
     fn check_persisted_uri_permission(&self, uri: &FileUri, mode: PersistableAccessMode) -> crate::Result<bool> {
-        impl_serde!(struct Req { uri: FileUri, mode: PersistableAccessMode });
-        impl_serde!(struct Res { value: bool });
-
-        let uri = uri.clone();
+        impl_se!(struct Req<'a> { uri: &'a FileUri, mode: PersistableAccessMode });
+        impl_de!(struct Res { value: bool });
 
         self.api
             .run_mobile_plugin::<Res>("checkPersistedUriPermission", Req { uri, mode })
@@ -261,8 +298,8 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
     
     fn get_all_persisted_uri_permissions(&self) -> crate::Result<impl Iterator<Item = PersistedUriPermission>> {
-        impl_serde!(struct Obj { uri: FileUri, r: bool, w: bool, d: bool });
-        impl_serde!(struct Res { items: Vec<Obj> });
+        impl_de!(struct Obj { uri: FileUri, r: bool, w: bool, d: bool });
+        impl_de!(struct Res { items: Vec<Obj> });
     
         self.api
             .run_mobile_plugin::<Res>("getAllPersistedUriPermissions", "")
@@ -280,10 +317,8 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
     
     fn release_persisted_uri_permission(&self, uri: &FileUri) -> crate::Result<()> {
-        impl_serde!(struct Req { uri: FileUri });
-        impl_serde!(struct Res;);
-
-        let uri = uri.clone();
+        impl_se!(struct Req<'a> { uri: &'a FileUri });
+        impl_de!(struct Res;);
 
         self.api
             .run_mobile_plugin::<Res>("releasePersistedUriPermission", Req { uri })
@@ -292,7 +327,7 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
     
     fn release_all_persisted_uri_permissions(&self) -> crate::Result<()> {
-        impl_serde!(struct Res);
+        impl_de!(struct Res);
 
         self.api
             .run_mobile_plugin::<Res>("releaseAllPersistedUriPermissions", "")
@@ -301,7 +336,7 @@ impl<R: Runtime> AndroidFs<R> for AndroidFsImpl<R> {
     }
 
     fn is_visual_media_dialog_available(&self) -> crate::Result<bool> {
-        impl_serde!(struct Res { value: bool });
+        impl_de!(struct Res { value: bool });
 
         self.api
             .run_mobile_plugin::<Res>("isVisualMediaDialogAvailable", "")
@@ -331,8 +366,8 @@ impl<R: Runtime> PublicStorage<R> for AndroidFsImpl<R> {
         mime_type: Option<&str>
     ) -> crate::Result<FileUri> {
 
-        impl_serde!(struct Req<'a> { dir: PublicDir, dir_type: &'a str });
-        impl_serde!(struct Res { name: String, uri: String });
+        impl_se!(struct Req<'a> { dir: PublicDir, dir_type: &'a str });
+        impl_de!(struct Res { name: String, uri: String });
 
         let dir = dir.into();
         let (_, dir_type) = match dir {
@@ -360,7 +395,7 @@ impl<R: Runtime> PublicStorage<R> for AndroidFsImpl<R> {
     }
 
     fn is_audiobooks_dir_available(&self) -> crate::Result<bool> {
-        impl_serde!(struct Res { value: bool });
+        impl_de!(struct Res { value: bool });
 
         self.api
             .run_mobile_plugin::<Res>("isAudiobooksDirAvailable", "")
@@ -369,7 +404,7 @@ impl<R: Runtime> PublicStorage<R> for AndroidFsImpl<R> {
     }
 
     fn is_recordings_dir_available(&self) -> crate::Result<bool> {
-        impl_serde!(struct Res { value: bool });
+        impl_de!(struct Res { value: bool });
 
         self.api
             .run_mobile_plugin::<Res>("isRecordingsDirAvailable", "")
@@ -385,7 +420,7 @@ impl<R: Runtime> PublicStorage<R> for AndroidFsImpl<R> {
 impl<R: Runtime> PrivateStorage<R> for AndroidFsImpl<R> {
 
     fn resolve_path(&self, dir: PrivateDir) -> crate::Result<std::path::PathBuf> {
-        impl_serde!(struct Paths { data: String, cache: String });
+        impl_de!(struct Paths { data: String, cache: String });
         
         static PATHS: std::sync::OnceLock<Paths> = std::sync::OnceLock::new();